backend/requirements.txt
# Updated Python dependencies for Flare integration
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
websockets>=12.0
pandas>=2.1.3
numpy>=1.26.2
aiohttp>=3.9.1
pydantic>=2.5.0

# Gemini AI
google-generativeai>=0.8.5

# Web3 / Ethereum (for Flare)
web3>=6.11.0
eth-account>=0.10.0
eth-utils>=2.3.0

# Flare-specific (if available)
# Note: Some Flare SDKs might not have Python versions yet
# We'll use web3.py to interact with Flare RPC

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
